+++ src/main/java/qna/domain/deleteHistory/ContentType.java
@@ -1,4 +1,4 @@
-package qna.domain;
+package qna.domain.deleteHistory;
 
 public enum ContentType {
     QUESTION, ANSWER;
