+++ src/main/java/qna/service/QnAService.java
@@ -6,7 +6,13 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import qna.CannotDeleteException;
 import qna.NotFoundException;
-import qna.domain.*;
+import qna.domain.deleteHistory.ContentType;
+import qna.domain.deleteHistory.DeleteHistory;
+import qna.domain.question.Question;
+import qna.domain.question.QuestionRepository;
+import qna.domain.question.answer.Answer;
+import qna.domain.question.answer.AnswerRepository;
+import qna.domain.user.User;
 
 import javax.annotation.Resource;
 import java.time.LocalDateTime;
