+++ src/test/java/qna/domain/question/QuestionTest.java
@@ -1,8 +1,9 @@
-package qna.domain;
+package qna.domain.question;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import qna.CannotDeleteException;
+import qna.domain.user.UserTest;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
