+++ src/test/java/qna/domain/question/answer/AnswerTest.java
@@ -1,8 +1,10 @@
-package qna.domain;
+package qna.domain.question.answer;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import qna.CannotDeleteException;
+import qna.domain.user.UserTest;
+import qna.domain.question.QuestionTest;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
